--- lib/Authentication/tempCodeRunnerFile.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-// class Login extends StatelessWidget {
-//   const Login({ Key? key }) : super(key: key);
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return SafeArea(
-//       child: ,
-//     );
-//   }
-// }
