+++ test/widget_test.dart
@@ -9,7 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:let_tutor/main.dart';
-import 'package:let_tutor/Authentication/login.dart';
+import 'package:let_tutor/Authentication/login/login.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
